restructure endpoint of GET / POST /_synapse/admin/v2/users/:user_id
