+++ src/lib.rs
@@ -4,5 +4,6 @@
 //! [api]: https://github.com/matrix-org/synapse/tree/master/docs/admin_api
 #![warn(missing_debug_implementations, missing_docs)]
 
+pub mod users;
 pub mod v1;
 pub mod v2;
