+++ src/users.rs
@@ -1,0 +1,5 @@
+//! Endpoints in the `/_synapse/admin/v<x>/users/` scope.
+
+mod common;
+pub mod create_or_modify;
+pub mod get_details;
